xX mod

+++ ft_printf/ft_printf.c
@@ -60,9 +60,9 @@ void	check_type(t_tag *tag)
 	if (*tag->fmt == 'd' || *tag->fmt == 'i')
 		tag->nbr = va_arg(tag->ap, int);
 	else if (*tag->fmt == 'u')
-		tag->nbr = va_arg(tag->ap, int);
+		tag->nbr = (unsigned int)va_arg(tag->ap, int);
 	else if (*tag->fmt == 'x' || *tag->fmt == 'X')
-		tag->nbr = va_arg(tag->ap, int);
+		tag->nbr = (unsigned int)va_arg(tag->ap, int);
 	digits_size(tag);
 }
